Favour the OSS dialect when hand-shaking
The W3C dialect is experimental and still under active development
and so should not be the preferred dialect to use.

To avoid being too demanding with memory usage, and to avoid
expensive operations such as laying profiles out on disk being run
more times than is strictly necessary, we now build the payloads
using StringBuilder.