Fix i18n typing tests.
Turns out that i18n is a tricky beast to get right. We were
incorrectly encoding keys that were more than a single byte
long as a character for the first byte, and a character for
the second. Clearly not right.

In addition, the JSON wire protocol and the W3C protocol
differ in how they want the values to be encoded. Delegate
sorting that difference down to the command codecs.