Handle older versions of the w3c protocol with a hack.
Until recently, the w3c protocol returned the value to
give to the user either in a "value" field, or as the
raw body of the response. The new version of the spec
now always returns the value as the "value" field.

Handle the older version of the protocol, spoken by
some version of geckodriver, by spamming in the raw
body of the response if there is no "value" field. This
is not technically correct, but it's good enough to
deal with this temporary state of affairs.