[rb] allow custom exec path, deprecate module methods
This also changes it so that the deprecation warning for reading driver_path is only shown if the user explicitly asks for it. Since it is called in the Service init, it would have always shown if there was no driver_path passed in.

Also removed some duplication when finding the binary

Signed-off-by: Titus Fortner <titusfortner@gmail.com>