ProtocolHandshake: be generous about duplicate keys
It's legit for a user to put the same key with
different values in the required and desired
capabilities. The expected behaviour is for the
required one to be chosen. Loosen the constraints
around this in the ProtocolHandshake by avoiding
the use of ImmutableMap and favouring the JRE's
built-in Collector.toMap instead.

Also removes a System.out.println from a test.