New Session: stream values from local to remote end
With the number of protocols we speak, large
capabilities (such as those that contain profile
objects), are a recipe for running out of RAM.
Try and mitigate this by streaming the capabilities
from the local end to the remote end.

Because of the way that we create the W3C
capabilities, it's entirely possible this won't
always work.