Fix the status handler
Turns out to be a nasty problem with instantiating some kinds
of handlers (those that don't take a parameter)

Interesting story: the lambda was created automatically by
a refactoring IDE. Each individual step was correct, except
the last one, which took a method call (`constructor.newInstance()`)
and turned into a method referened (`constructor::newInstance`).

When called, the original version of this didn't pass any
arguments, and everything was fine. The mathod reference version,
however, caused the sessionId to be passed to the `newInstance`
method as argument, causing grief. Still. Fixed now.