Replace backslashes in paths "correctly" on UNIX
The FirefoxOptions attempts to standardise paths to use
forward slashes since these are most portable. The
existing code works on Windows, but not a UNIX-like OS,
which is a pity.

The fix here will obviously fail if given a UNIX path
with a valid backslash in it, but we'll assume that
never happens.