Make Content-Type match spec requirements
I guess, technically, JSON can be thought of as a
kind of Javascript (after all, it's JavaScript
Object Notation), but the spec says that we should
be using `application/json`, which feels more
correct too.

Fixes a problem where the ruby remote driver did
not want to start a new session.