Switch Hub to use the NewSessionPayload to find Capabilities
It would be better if we passed around the NewSessionPayload
itself, but given that it's Grid, we pull the first set of
capabilities from the payload.

Because the NewSessionPayload prefers to use the w3c capabilities
we need to make sure that the test uses spec-compliant markers
to indicate that options are important.