Adding ability to detect spec-compliant capabilities in .NET
This commit updates the .NET `DesiredCapabilities` class to allow the
language bindings to construct the correct type of payload for the new
session command before sending the command to the remote end. This will
remove the burden from intermediate nodes (like the Java standalone server
and grid) from incorrectly parsing a driver using the legacy protocol
dialect (Edge, Safari, Chrome) and throwing an error on session creation.

Additionally, this commit also deprecates the static methods on the
`DesiredCapability` classes for specific browsers in .NET. For over a
year, the guidance for .NET users has been to use a browser-specific
options class (`FirefoxOptions`, `InternetExplorerOptions`, etc.) to
set specific capabilities for the driver to be instaniated. To use the
options classes with the Java standalone remote server or grid, call
the `ToCapabilities` method on the options object.

Fixes issue #4443.