Updating .NET FirefoxOptions to not force specifying binary location
In the Marionette/Geckodriver case, the service executable can locate the
Firefox executable on its own when looking in default installation
locations. There is no need for the language bindings to go spelunking
through the registry in that case any longer because the bindings do not
have to directly launch the Firefox executable.