[js] Update examples & docs to not rely on btnG on the Google search page
After years of stability, the Google search page has recently changed so
that btnG is no longer visible on the page. This causes all of our
examples that interact with it to fail. Instead of trying to click the
search button, change examples to simply press RETURN after typing the
query.

Fixes #4491