Handle permission denied error upon attempt to start multiple Edge drivers
This only happens on Windows and Edge and we should consider the
port owned when getting such an error.

Potentially, this might lead to a worse case if user is not allowed to
bind ports at all. In this case, we will end up looking for a port
forever. Let's give it a shot anyway and see if it causes any problems
for users.