[js] Remove the firefox.Binary class.
This is more cleanup from removing support for the legacy FirefoxDriver. The
only functionality it still provided was the ability to specify the path to a
specific binary and define additional command line arguments for when starting
Firefox. Both of these options can be configured using the firefox.Options
class.