Remove quirky platform handling from ImmutableCapabilities
If you called the constructor for `ImmutableCapabilities` that
took a `Map`, it would "magically" try and figure out what the
value for both `platform` and `platformName` should be. No other
version of `Capabilities` did this, including
`ImmutableCapbilities` itself.

Rather than try and patch around this, make `getPlatform` do the
hard work, and make the constructor do less work. This can only
be a Good Thing.