updated references to 'btnG' (deprecated) to 'btnK'
A number of legacy references to btnG exist within a variety of files throughout the repo. These have been replaced with current name of search button field on Google home page ('btnK)

Could not raise issue through GitHub, but substance of it is this: Issue relates to legacy reference to 'btnG' which I gather used to be the name property assigned to the input for 'Google Search' HTML button used as an example throughout the repo. This has since updated to 'btnK' on Google's home page. Comments and examples need to be updated to reflect the change, or otherwise modified.

Signed-off-by: Alex Rodionov <p0deje@gmail.com>