Move NewSessionPayload so it can be used by ProtocolHandshake
There's a lot of overlap between the ProtocolHandshake and
the NewSessionPayload. Time to deal with this overlap between
the two.

The main change is to make the NewSessionPayload capable of
writing to an Appendable. It does so by streaming the
Capabilities used to create the payload as the OSS capabilities
and those used by old geckodrivers, and then streaming every
capability through transforms to generate spec compliant
capabilities.

There appears to be a bug where we always generate synthetic
capabilities, but we can deal with that later.