[js] Removed various browsers from the list of "standard" browser names supported by selenium-webdriver.
ANDROID: there hasn't been an AndroidDriver in a long time. Users should instead
use the chromedriver configured for Chrome on Android.

IPAD & IPHONE: there hasn't been meaningful support for either of these in a
a long time. Surfacing them in the API is misleading to users.

OPERA: The operadriver appears to be abandoned. Since the operadriver and Opera
are just forks of chromedriver and Chromium, respectively, users can get by
using Chrome.

PHANTOM_JS & HTMLUNIT: Users should be using Chrome or Firefox in headless mode