Return NoSuchElement error in IE with incorrect internal IDs
The IE driver, to be spec compliant, must return "no such element" if an
invalid internal WebDriver ID is passed in the URL. Previously, the driver
would assume that the ID was one that was previously provided by the
driver, but now has gone stale. This may cause discomfort for some rare
use cases where the incorrect exception is being looked for.