Reverting "Dropping capabilities that can cause geckodriver to reject a New Session request"
For "browserName", as geckodriver is designed for Firefox, it's
right for geckodriver to complain if the browser name indicates
something else. The solution here is not to set the browser name
at all (which can be done by setting the capability to `null`)

For browser version, the spec says that the matching is done in
an implementation specific way. It's probably best to leave that
out and configure the Grid to hand out nodes with the right
version for you.

`platformName` is more contentious. There's a note in the spec that
recommends `Platform` families be respected, and there's an open
issue for geckodriver about this:

https://bugzilla.mozilla.org/show_bug.cgi?id=1421233

This reverts commit b90d793ea824275ac3c79cbed5f8c41b1196e4a4.