Allow the default HttpFactory to be switched using a system property
Set `webdriver.http.factory` to `okhttp` to use the version backed
by OkHttp, or `apache` to use the version backed by Apache's
HttpClient. By default, the OkHttp implementation is used, but
this allows users to back out.

This system property is short-lived, and will be removed once we're
confident that the OkHttpClient works as well as the Apache one.