Exposing HostName property for .NET DriverService class
Previously, the host name for a DriverService was hard-coded to be
"localhost." This is normally acceptable, as a DriverService is only
intended to be used locally, and most executables that are represented by
a DriverService (chromedriver, geckodriver, IEDriverServer.exe, etc.) do
not even allow connections from external machines. However, there is a
legitimate use case (.NET Standard 2.0) where a local IP address
("127.0.0.1" or "::1") is preferrable to using the "localhost" moniker.
This HostName property allows the user to specify that IP address. Note
that the property defaults to "localhost," preserving previous behavior.