Making JavaScript execution asynchronous for executeScript in IE
Prior to this change, the execution of JavaScript via the executeScript
method was synchronous. This would mean that if the script passed into the
driver called the JavaScript `alert()`, `confirm()`, or `prompt()`
functions, the WebDriver code would hang until the dialog was manually
dismissed. Moving this to asynchronous execution allows the driver to be
compliant with the W3C WebDriver Specification, that mandates that user
prompts triggered by executeScript should be able to be handled by the
WebDriver alert handling functions. However, this spec compliance comes at
several costs.

The first cost is complexity. The IE driver uses COM interfaces to
communicate with Internet Explorer. COM is inherently single-threaded, and
requires a fair amount of manipulation to allow accessing the objects on
separate threads. This leads to requiring a fairly complex system of
communication and synchronization between the threads to give the illusion
of a synchronous script execution.

The second cost is performance. Because of the need for synchronization,
this change **may** introduce a performance hit in the IE driver. This is
unfortunate, but unavoidable. The benefit of having user-initiated
JavaScript not block and hang WebDriver code far outweighs whatever
relatively small performance hit has been introduced.