Updating IE driver to better handle non-US keyboards in sendKeys
In the transition to support the W3C specification, using sendKeys on
non-US keyboard layouts got broken. This is an attempt to resolve that
issue. Previously, we were only checking a high-order byte for the need to
press the shift key, and not other modifier keys (control or alt) in
sending keystrokes. This behavior has been restored. Fixes issue #4253.