Support binding to IPv6 loopback in IE driver
When no host is specified on the command line, and no list of whitelisted
IP addresses are specified, IEDriverServer.exe will now bind to both the
IPv4 and IPv6 loopback adapters. If no IPv6 stack is found on the OS, the
executable will retry binding only to IPv4. Note that when using
whitelisted IP addresses, only IPv4 is supported. This is a limitation of
the Civetweb HTTP server upon which IEDriverServer.exe relies. This commit
also updates the version of Civetweb to 1.10.