Refactoring user input simulation code in IE driver
This commit breaks out the three modes of user input simulation (native
events using SendMessage, native events using SendInput, simulated events
using JavaScript) into separate classes, making it easier to debug.
Additionally, this commit removes the code for enabling persistent hovers.
This code had been inadvertently disabled in a separate refactor some
months ago, and there have been no issue reports in that area, leading the
IE driver development team to believe that the issue in IE that required
persistent hovers has been resolved (at least in IE11). The code for
persistent hovers has been retained in a new, separate class in the
project, but this class is not built as part of the IE driver at this
time. If issue reports warrent its inclusion in the future, it can be
easily re-added.