Fixing IE cookie retrieval for IE 10 and 11
Internet Explorer relies on a Windows component called WinINet to
communicate over the Internet. This includes WinINet APIs for getting and
setting cookies. Prior to Windows 10, persistent cookies were stored in
files on the hard drive. A change in WinINet moved the persistent cookie
storage from files on the disk to an Extensible Storage Engine (ESE or
"Jet Blue") database, which rendered them unreadable by the IE driver.
This commit restores the functionality of getting cookies by using an API
function called InternetGetCookieEx2.

This API function was introduced in the Windows 8.1 SDK, which means that
it won't be available to older versions of WinINet. In that case, we fall
back to the original behavior of attempting to parse cookie files from the
disk. There is a chance that the detection of the availability of this API
is flawed in the IE driver.