Marking .NET ExpectedConditions obsolete
This implementation of PageFactory for .NET is deeply flawed.
Additionally, using the ExpectedConditions provides no benefit over
directly using lambda functions (anonymous methods) directly in
one's code. Since the community appears to believe that an "official"
repository of wait conditions is desireable, the existing code has
been migrated to a new repository under a new organization on GitHub
(https://github.com/DotNetSeleniumTools/DotNetSeleniumExtras). It is
hoped that this will encourage a volunteer from the community to take
ownership of this code. Users should update their references and
migrate their code to use `SeleniumExtras.ExpectedConditions`. This
implementation will be removed from the .NET language bindings in a
future release.