Removing direct call to InternetGetCookieEx2 API in IE
Versions of IE prior to 11 do not have access to this API. Linking
directly to it via the Windows 8.1 SDK creates an import entry in the
IEDriver.dll library used by the driver. So that the driver can continue
to be used against earlier releases, we will call the API using
GetModuleHandle and GetProcAddress only if it exists in the version of
WinINet being used by the browser. Fixes issue #5603.