Adding Protected Mode boundary closing mechanism in IE
When a Protected Mode boundary is crossed (entering or exiting Protected
Mode), the existing browser instance is destroyed and a new one created in
its place by Interent Explorer. This commit adds detection for when a
browser instance is being closed, but without an explicit call to the
WebDriver close() or quit() methods. When this is detected, it's likely a
Protected Mode boundary is being crossed, and all subsequent commands in
the WebDriver session will fail. In this case, the driver will now write
to the log that the browser has been asked to exit without the user
explicitly requesting it. The detection is not perfect, since it's
possible to legitimately click a link that closes the browser window, and
this is indistinguishable from clicking a link that navigates to a URL
that causes a Protected Mode boundary crossing. Nevertheless, this logging
is being added so that users can see what may be happening when they
receive errors like "Unable to get current browser."