Update Http.request in default.rb to include a delay in retries
Update Http.request to include a sleep between retries for cases where Errno::ECONNABORTED, Errno::ECONNRESET, Errno::EADDRINUSE occur.   Without this sleep the retries happen immediately and do not give the remote host and/or connection to it, any time to resolve the connection issue.  This results in seemingly random connection problems and test failures when using Selenium Webdriver in a grid setup with 1000s of connections running simultaneously.

Signed-off-by: Alex Rodionov <p0deje@gmail.com>