Fixing .NET HttpCommandExecutor event for RemoteWebDriver
This commit moves the firing of the event raised before sending a remote
HTTP request to before the setting of the request body. This allows the
request settings to be properly set without throwing an exeception. This
commit also renames the event and event args object to
`SendingReomteHttpRequest` and `SendingRemoteHttpRequestEventArgs`,
respectively. The previous event and event args are marked as deprecated
for removal in a future release, and the previously named event is now a
no-op, and will never be fired, even if a handler is attached.