Adding property to force use of legacy protocol for Safari in .NET
Starting with Safari Technology Preview 58, and continuing with the driver
released with Safari 12, the driver executable uses the W3C Specification
dialect of the prototcol by default. It includes a command-line switch,
however, to force the use of the legacy open-source project dialect of the
protocol. This commit adds a property to the SafariDriverService object
allowing users to set that command-line switch. Note that if this property
(false by default) is set to true for a version of the driver that does
not support the switch, it will cause errors. If set to false, the .NET
bindings do not modify the safaridriver command line at all.