Refactoring .NET capability handling
This commit introduces a large, and intrusive refactor of the .NET
bindings' capability handling. It introduces a new class,
`RemoteSessionSettings`, for use with creating remote sessions via
`RemoteWebDriver`. Additionally, the `DesiredCapabilities` class is now
marked as deprecated and will generate a compile warning on its use.

The RemoteSessionSettings class is designed for use cases where the user
wants to create a remote session using RemoteWebDriver where all of the
nodes in the session creation support the W3C WebDriver Specification
dialect of the wire protocol. This class is designed to be used in
conjunction with the browser-specific driver options classes for matching
capabilities on the remote end.

For single-browser cases, it is still possible (though unnecessary) to
use a browser-specific driver options class (`ChromeOptions`,
`FirefoxOptions`, etc.) by calling the `ToCapabilities()` method, but
the capabilities returned are now read-only, and cannot be added to.
Users who feel the need to modify the options class after converting to
capabilities are encouraged to add the additional capabilites to the
options class before the conversion by appropriate use of the
`AddAdditionalCapability` method.