Adding wait for navigation to IE accept and dismiss alert commands
When accepting or dismissing an alert, particularly when it is raised
during the onload event, it is possible for the driver to not recognize
that navigation is still pending once the alert is handled. This change is
an attempt at better allowing the driver to recognize this situation and
act accordingly.