Exposing additional extension points in .NET bindings
For the .NET implementation of HttpCommandExecutor, we now expose a Proxy
property for proxying requests between the .NET code and the remote end
WebDriver implementation. Note carefully that this is entirely separate
from the proxy used by the browser for monitoring communications between
the browser and the server hosting the page being automated.

Additionally, this change exposes the IsKeepAliveEnabled property to make
it easier to set whether or not the "keep alive" header is sent when
communicating between the language bindings and the remote end WebDriver
implementation.

Finally, for local use, the DriverServiceCommandExecutor now exposes a
property to allow the user to set tho newly-exposed properties by allowing
access to the internal HttpCommandExecutor of the driver service.