Introduce basic JPMS support
Because we build with and target java 8, we can't use
the normal toolchain to compile `module-info` classes.
One solution to this is to ask everyone to install the
latest java toolchain and use that, but java 10 (which
is the latest) isn't an LTS, so will go pretty soon.

Instead, we use `javaparser` to read `module-info.java`
files and then use the latest version of `byte buddy`
to emit the correct class.

To allow simple globbing of java classes, I've chosen
to use a `.txt` file name ending for the module-info
files.

Once Java 11 ships and Buck supports it, we'll rejig
the toolchain to remove this particular hack.