Move all option processing on boot into one place.
This isn't particularly elegant, and there's a lot of
duplication, but this means as we slide the new config
into place we don't need to hunt down quite so much
stuff.

It's deeply unfortunate that we expose jcommander.
My hope is that once we clean everything up, we can
go back and hide our use of jcommander once and for
all.