Updating IE driver for the case of clicking element with children
The case where you attempt to click on an element that must be scrolled
into view, and where the element has child elements, but the first
node is a text node was broken for IE. In this case, to guarantee the
proper element is clicked (and that an embedded child does not receive
the click), the driver would attempt to click on the text node. However,
the coordinate calculation when the element had to be scrolled into view
was flawed. This commit changes from using the boundingLeft and
boundingTop properties of IHTMLTextRangeMetrics to using offsetLeft and
offsetTop in an attempt to correct the calculation of the location of
the text to be clicked on.