Making JSON serialization of JavaScript objects more compliant
When returning values from user-specified JavaScript, the presence of a
toJSON() property indicates that function is what should be used to
serialize the object. Additionally, for objects that are not recognized
as JavaScript objects via the properties of the IDispatch, and do not
have a toJSON property, the driver will now make a last-ditch attempt
to serialize the value for return by calling JSON.stringify().