Introduce a Node for running sessions
The Node models a place where an individual webdriver
session is running. A node can run more than one session
at a time.

Again, this is implemented in three pieces: an abstract
base class, a "remote session map", which takes a simple
request and converts it to the http calls specified in
the base class, and an implementation that is designed
to run locally.