[bazel] Restructure build files
The following changes have been made:

* Following bazel idioms

Whereas Buck required the target to be fully qualified, the
idiom in Bazel is to omit the specific target within a build
file provided that target is the name of the directory. That
is `//third_party/java/guava:guava` is the same as
`//third_party/java/guava`. Bazel favours brevity.

* Minimal visibility

Unpicking dependency graphs is a PITA. To avoid accidentally
making life hard for ourselves, all targets default to bring
private. We only expose a target if it is absolutely required.
Public visibility should be vanishingly rare in our build
files.

* One visible rule per build file

Other than breaking cycles in dependency graphs, we should be
striving for a single visible target per build file. While
this is not a hard-and-fast rule, it makes wiring things up
simpler. Typically, this single rule will export anything
that should be visible from the file.

* Follow seams created by deployable artifacts

In the java world, the deployable artifacts are maven jars.
By definition, these are public, but for the sake of keeping
things under control, we may limit the visibility of them.
Despite this, if a library is part of a deployable artifact,
then rather than allowing people to depend on the library
itself, we should be depending on the deployable artifact.
This is why so many targets depend on the remote jar.