Implementing create new window commmand
The W3C WebDriver Specification recently added the ability to create
a new top-level browsing context, or what appears to the user to be a
new tab or window. This change implements that command. Though the
command payload contains a hint argument as to what type of new top-
level browsing context the user desires (tab or window), the driver
implementation is free to ignore this type hint if it does not
support the type being requested. Since the IE driver has never
supported, and still does not support, creation of new top-level
browsing contexts using a new tab, in accordance with the spec, the
driver will create the new context in a new window, even if a new
tab is requested. Please note this is not a bug; it is as designed,
and is still fully compliant with the specification with this
behavior.